Refactoring to use more queueDialog
